+++ tests/conftest.py
@@ -35,6 +35,16 @@ MOODBOT_MODEL_PATH = "examples/moodbot/models/dialogue"
 
 DEFAULT_ENDPOINTS_FILE = "data/test_endpoints/example_endpoints.yml"
 
+TEST_DIALOGUES = ['data/test_dialogues/default.json',
+                  'data/test_dialogues/formbot.json',
+                  'data/test_dialogues/moodbot.json',
+                  'data/test_dialogues/restaurantbot.json']
+
+EXAMPLE_DOMAINS = [DEFAULT_DOMAIN_PATH,
+                   "examples/formbot/domain.yml",
+                   "examples/moodbot/domain.yml",
+                   "examples/restaurantbot/restaurant_domain.yml"]
+
 
 class CustomSlot(Slot):
     def as_feature(self):