+++ tests/test_trackers.py
@@ -1,6 +1,7 @@
-import fakeredis
 import glob
 import json
+
+import fakeredis
 import pytest
 
 from rasa_core import training, restore
@@ -11,15 +12,15 @@ from rasa_core.events import (
     UserUttered, ActionExecuted, Restarted, ActionReverted,
     UserUtteranceReverted)
 from rasa_core.tracker_store import InMemoryTrackerStore, RedisTrackerStore
-from rasa_core.tracker_store import (
-    TrackerStore)
+from rasa_core.tracker_store import TrackerStore
 from rasa_core.trackers import DialogueStateTracker, EventVerbosity
-from tests.conftest import DEFAULT_STORIES_FILE
+from tests.conftest import (DEFAULT_STORIES_FILE,
+                            EXAMPLE_DOMAINS, TEST_DIALOGUES)
 from tests.utilities import (
     tracker_from_dialogue_file, read_dialogue_file,
     user_uttered, get_tracker)
 
-domain = Domain.load("data/test_domains/default.yml")
+domain = Domain.load("examples/moodbot/domain.yml")
 
 
 @pytest.fixture(scope="module")
@@ -46,7 +47,7 @@ def stores_to_be_tested_ids():
 
 
 def test_tracker_duplicate():
-    filename = "data/test_dialogues/inform_no_change.json"
+    filename = "data/test_dialogues/moodbot.json"
     dialogue = read_dialogue_file(filename)
     tracker = DialogueStateTracker(dialogue.name, domain.slots)
     tracker.recreate_from_dialogue(dialogue)
@@ -90,32 +91,34 @@ def test_tracker_store_storage_and_retrieval(store):
 
 @pytest.mark.parametrize("store", stores_to_be_tested(),
                          ids=stores_to_be_tested_ids())
-@pytest.mark.parametrize("filename", glob.glob('data/test_dialogues/*json'))
-def test_tracker_store(filename, store):
+@pytest.mark.parametrize("pair", zip(TEST_DIALOGUES, EXAMPLE_DOMAINS))
+def test_tracker_store(store, pair):
+    filename, domainpath = pair
+    domain = Domain.load(domainpath)
     tracker = tracker_from_dialogue_file(filename, domain)
     store.save(tracker)
     restored = store.retrieve(tracker.sender_id)
     assert restored == tracker
 
 
-async def test_tracker_write_to_story(tmpdir, default_domain):
+async def test_tracker_write_to_story(tmpdir, moodbot_domain):
     tracker = tracker_from_dialogue_file(
-        "data/test_dialogues/enter_name.json", default_domain)
+        "data/test_dialogues/moodbot.json", moodbot_domain)
     p = tmpdir.join("export.md")
     tracker.export_stories_to_file(p.strpath)
     trackers = await training.load_data(
         p.strpath,
-        default_domain,
+        moodbot_domain,
         use_story_concatenation=False,
         tracker_limit=1000,
         remove_duplicates=False
     )
     assert len(trackers) == 1
     recovered = trackers[0]
-    assert len(recovered.events) == 7
-    assert recovered.events[5].type_name == "slot"
-    assert recovered.events[5].key == "name"
-    assert recovered.events[5].value == "holger"
+    assert len(recovered.events) == 11
+    assert recovered.events[4].type_name == "user"
+    assert recovered.events[4].intent == {'confidence': 1.0,
+                                          'name': 'mood_unhappy'}
 
 
 async def test_tracker_state_regression_without_bot_utterance(default_agent):