+++ CHANGELOG.rst
@@ -41,6 +41,8 @@ Changed
 - made ``message_id`` an additional argument to ``interpreter.parse``
 - changed payloads from "text" to "message" in files: server.yml, docs/connectors.rst, 
   rasa_core/server.py, rasa_core/training/interactive.py, tests/test_interactive.py
+- dialogue files in ``/data/test_dialogues`` were updated with conversations
+  from the bots in ``/examples``
 
 Removed
 -------